Version 1.0
Used to periodically check whether the content in the keyword list exists on the query page. If keywords are found, an email is automatically sent via Outlook. 
Only applicable to the URL proposed by Jürgen. Replacing it with another URL will cause the automatic acceptance of cookies to fail.